authentification added for edit delete

+++ controllers/feed.js
@@ -3,6 +3,7 @@ const {Post} = require("../models/post");
 const path = require("path");
 const {publicImagesPath} = require("../constants/publicImages");
 const fs = require('fs/promises')
+const {User} = require("../models/user");
 
 const fetchPost = async (req, res, next) => {
   const postId = req.params.postId
@@ -34,7 +35,7 @@ const getPosts = async (req, res, next) => {
 
 const getPostById = async (req, res, next) => {
   try {
-    const post =  req.post
+    const post = req.post
 
     return res.status(200).json({
       post
@@ -52,21 +53,26 @@ const createPost = async (req, res, next) => {
     throw error
   }
   const {title, content} = req.body
-  const post = new Post({title, content, creator: {name: 'Dima'}, imageUrl: req?.file?.filename})
+  const post = new Post({title, content, imageUrl: req?.file?.filename, creator: req.userId})
   try {
     await post.save()
+    const user = await User.findById(req.userId)
+    if (!user) {
+      throw new Error('No user found')
+    }
+    user.posts.push(post)
+    await user.save()
     return res.status(201).json({
       message: 'created',
       post: {
         _id: new Date(),
         title,
         content,
-        creator: {name: 'Dima'},
+        creator: user,
         createdAt: new Date()
       }
     })
-  }
-  catch (e) {
+  } catch (e) {
     next(e)
   }
 }
@@ -81,6 +87,11 @@ const updatePost = async (req, res, next) => {
   const imageUrl = req.file?.filename
   try {
     const post = req.post
+    if (post.creator.toString() !== req.userId) {
+      const e = new Error('Deleting not allowed')
+      e.statusCode = 403
+      throw e
+    }
     if (!post) {
       const e = new Error('Post not found')
       e.statusCode = 404
@@ -98,8 +109,7 @@ const updatePost = async (req, res, next) => {
     await post.save()
 
     return res.status(200).json({post})
-  }
-  catch (e) {
+  } catch (e) {
     next(e)
   }
 }
@@ -107,14 +117,19 @@ const updatePost = async (req, res, next) => {
 const deletePost = async (req, res, next) => {
   const post = req.post
   try {
+    if (post.creator.toString() !== req.userId) {
+      const e = new Error('Deleting not allowed')
+      e.statusCode = 403
+      throw e
+    }
+
     await deleteImage(post.imageUrl)
     await Post.findByIdAndRemove(post._id)
     return res.status(200).json({
       message: 'deleted'
     })
-  }
-  catch (e) {
-      next(e)
+  } catch (e) {
+    next(e)
   }
 
 }
@@ -122,7 +137,9 @@ const deletePost = async (req, res, next) => {
 const deleteImage = async (imageUrl) => {
   try {
     await fs.unlink(path.join(process.cwd(), 'public', 'images', imageUrl))
-  } catch (e) {console.error(e)}
+  } catch (e) {
+    console.error(e)
+  }
 }
 
 module.exports = {
